ref: move password file to use database dir
